flappyBird
This is flappy bird game. The bird is constantly moving forward and falls down when you do not touch the screen, and rises when you click on the screen. The aim is to pass the bird between the two pipes. I made this application using Kotlin language on Android Studio.